add TeamDto, StadiumDto, ScorerDto, LineupDto & CoachDto

+++ BackEnd-MatchesService/src/main/java/tn/iit/entity/Scorer.java
@@ -4,12 +4,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
-import java.math.BigInteger;
-
 @Data
 @Builder
 @Jacksonized
 public class Scorer {
     private Player playerScoring;
-    private BigInteger scoringTime;
+    private int scoringTime;
 }
